Fixing inheritance/python import

+++ tools/cpp/KML/include/IStreaming.h
@@ -20,12 +20,12 @@ namespace KML
     class IStreaming
     {
         public:
-            /**
+            /*!
              * @brief   IStreaming Constructor.
              * @details Constructor for every Streaming algorithms.
              * @param   windowSize Window size of number of items held in memory (range [0,..)).
              */
-            IStreaming(const uint64_t& windowSize = 0);
+            IStreaming(const size_t& windowSize = 0);
 
             /*!
              * @brief Default Copy Constructor.
@@ -37,51 +37,51 @@ namespace KML
              */
             IStreaming& operator=(const IStreaming& a_rhs) = delete;
 
-            /**
+            /*!
              * @brief Destructor.
              */
             virtual ~IStreaming();
 
-            /**
+            /*!
              * @brief  Takes a single observation and updates the deriving class.
              * @param  observation The observation used to update deriving class.
              */
             virtual void update(const T& observation) = 0;
 
-            /**
+            /*!
              * @brief  Takes in a list of one or more observations to update 
              *         the deriving class.
              * @param  observations List of observations to update deriving class.
              */
             virtual void update(const std::vector<T>& observations);
 
-            /**
+            /*!
              * @brief  Evaluate everything in the current window.
              * @return The deriving class evaluate. 
              */
             virtual double evaluate() const = 0;
 
-            /**
+            /*!
              * @brief Retrive the current window.
              * @return Pointer to current window.
              */
             virtual std::deque<T>* getWindow();
 
         protected:
-            /**
+            /*!
              * @brief Pointer to current window of observations.
              */
             std::deque<T>* m_window;
 
-            /**
+            /*!
              * @brief The current window size.
              */
-            const uint64_t m_windowSize;
+            const size_t m_windowSize;
 
-            /**
+            /*!
              * @brief The number of observations seen by deriving class.
              */
-            uint64_t m_historyCount;
+            size_t m_historyCount;
     };
 } // KML
